Add libfabric memory region/registration code
This is taken from the original HPX/GHEX implementation
and contains some remnants that can be removed once testing
is complete and all is working.

This version uses a memory region that stores base address
offset, size and used space which may not all be needed in
a final version.