Add a function to register user memory
Sometimes the user will want to register memory that is not owned
by the heap/allocator. This provides a function that returns a
pointer compatible to a normal allocation from a user registration.

Add flags to the registration function so that transport layers
can use them to modify their behaviour