Remove heap instance singleton that was added for hpx support
A work around for this has been found that removes the need for the
singleton. One problem with using a (static) singleton is that destruction
takes place at shutdown, when needed handles to fabric APIs might have
already been released, so it is better to managed alloc/dealloc of pools
with manual checks in user code.